// src/api/generated/models/staffListSortOrder.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Bookora API
 * Contract OpenAPI chính thức cho backend Bookora. Frontend dùng cookie credentials, cookie accessToken/refreshToken và header X-CSRF-Token cho mutation cần CSRF.
 * OpenAPI spec version: 1.0
 */

export type StaffListSortOrder = typeof StaffListSortOrder[keyof typeof StaffListSortOrder];


export const StaffListSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;